# contracts/openapi/alignment-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Traffic Analysis Platform Alignment API",
    "version": "1.2.0",
    "description": "Additive contract for the four remediation vertical slices. Existing APIs remain available during migration."
  },
  "servers": [{"url": "/api"}],
  "paths": {
...
    "/v1/alerts/batches/assign/{batch_id}": {
      "get": {
        "operationId": "getAlertBatchAssignment",
        "x-required-scope": "alert:write",
        "x-feature-id": "F-ALERT-004",
        "x-rollout-flag": "ALERT_BATCH_ASSIGNMENT_V1_ENABLED",
        "parameters": [{"name": "batch_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}],
        "responses": {
          "200": {"description": "Tenant-bound batch and complete ordered per-item result set; accepted and running states remain explicitly partial", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Envelope"}}}},
          "400": {"description": "Invalid batch identifier"},
          "401": {"description": "Authenticated tenant and user required"},
          "403": {"description": "alert:write scope required"},
          "404": {"description": "Batch is absent from the authenticated tenant"},
          "503": {"description": "Feature disabled or PostgreSQL schema unavailable"}
        }
      }
    },
...
    "/v1/alerts/batches/assign/{batch_id}/compensations/{request_id}": {
      "get": {
        "operationId": "getAlertBatchAssignmentCompensation",
        "x-required-scope": "alert:write",
        "x-feature-id": "F-ALERT-004",
        "x-rollout-flag": "ALERT_BATCH_ASSIGNMENT_COMPENSATION_V1_ENABLED",
        "parameters": [
          {"name": "batch_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
          {"name": "request_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}
        ],
        "responses": {
          "200": {"description": "Tenant-bound compensation aggregate and complete ordered per-item terminal or in-progress receipts", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Envelope"}}}},
          "400": {"description": "Invalid batch or compensation request identifier"},
          "401": {"description": "Authenticated tenant and user required"},
          "403": {"description": "alert:write scope required"},
          "404": {"description": "Compensation is absent from the authenticated tenant"},
          "503": {"description": "Compensation feature disabled or PostgreSQL schema unavailable"}
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
...
      "Envelope": {
        "type": "object",
        "required": ["data", "meta", "error"],
        "properties": {
          "data": {},
          "meta": {"$ref": "#/components/schemas/Meta"},
          "error": {"oneOf": [{"type": "null"}, {"$ref": "#/components/schemas/Error"}]}
        }
      },
      "Meta": {
        "type": "object",
        "required": ["contract_version", "schema_version", "snapshot_id", "as_of", "generated_at", "trace_id", "result_code", "partial", "missing_sections", "source_watermarks"],
        "properties": {
          "contract_version": {"type": "integer"},
          "schema_version": {"type": "integer"},
          "snapshot_id": {"type": "string"},
          "as_of": {"type": "string", "format": "date-time"},
          "generated_at": {"type": "string", "format": "date-time"},
          "trace_id": {"type": "string"},
          "result_code": {"type": "string"},
          "operation_id": {"type": "string"},
          "tenant_id": {"type": "string"},
          "partial": {"type": "boolean"},
          "missing_sections": {"type": "array", "items": {"type": "string"}},
          "source_watermarks": {"type": "object", "additionalProperties": {"type": "string"}},
          "projection_status": {"type": "string"}
        }
      },
      "Error": {
        "type": "object",
        "required": ["code", "message", "trace_id", "retryable"],
        "properties": {
          "code": {"type": "string"},
          "message": {"type": "string"},
          "trace_id": {"type": "string"},
          "retryable": {"type": "boolean"},
          "field_errors": {"type": "array", "items": {"type": "object", "required": ["field", "code", "message"], "properties": {"field": {"type": "string"}, "code": {"type": "string"}, "message": {"type": "string"}}}},
          "operation_id": {"type": "string"},
          "current_revision": {"type": "integer"},
          "projection_status": {"type": "string"},
          "status_url": {"type": "string"},
          "expected_consistency_window": {"type": "string"}
        }
      },
...
    }
  }
}